pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "cv-agent"
version = "0.1.0"
description = "CV discovery pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.14.0",
    "agent-framework>=1.0.0",
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "langchain-community>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=5.0.0",
    "qdrant-client>=1.17.1",
    "tenacity>=8.2.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]

[project.scripts]
cv-agent-serve = "cv_agent.__main__:main"
cv-reindex = "cv_agent.jobs.reindex:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_agent"]

[tool.ruff]
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
